Revert "WAQ refactor[WIP] (#1496)"
This reverts commit 83a4569363882867d0719e2c10ea57c2689efaa0.
