--- neural_compressor/adaptor/torch_utils/waq/README.md
@@ -1,1 +1,0 @@
-weight and activation quantization algorithm
