+++ test/quantization/test_weight_only_quantization.py
@@ -6,7 +6,7 @@ import torch
 import transformers
 
 from neural_compressor.adaptor.torch_utils.model_wrapper import WeightOnlyLinear
-from neural_compressor.adaptor.torch_utils.waq import GraphTrace
+from neural_compressor.adaptor.torch_utils.smooth_quant import GraphTrace
 from neural_compressor.adaptor.torch_utils.weight_only import (
     autoround_quantize,
     awq_quantize,
